refactor: centralize random helpers
